\name{OSTE-package}
\alias{OSTE-package}
\docType{package}
\title{
Optimal Survival Trees Ensembles}
\description{
This package consists of function for growing survival trees ensemble, that are grown by the method of random survival forest. The survival trees grown are assessed for both individual and collective performances. The ensemble can give promising results on fewer survival trees selected based on their individual and collective performance in the final ensemble.
}
\details{
\tabular{ll}{
Package: \tab OSTE\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2021-11-07\cr
License: \tab GPL (>= 3.5.0)\cr
}

}
\author{ Naz Gul, Nosheen Faiz, Zardad Khan and Berthold Lausen.

Maintainer: Naz Gul <nazgul@awkum.edu.pk>
}
\references{
Gul, N., Faiz, N., Brawn, D., Kulakowski, R., Khan, Z., & Lausen, B. (2020). Optimal survival trees ensemble. arXiv preprint arXiv:2005.09043.
}
\keyword{OSTE}
